mach-o: Don't split version into a few fields
X.Y.Z tuple is used for minos version and sdk version fields in a number of
load commands:
  LC_VERSION_MIN_MACOSX
  LC_VERSION_MIN_IOS
  LC_VERSION_MIN_WATCHOS
  LC_VERSION_MIN_TVOS
  LC_BUILD_VERSION

We could use a macro to avoid code duplication for reading and setting
the X/Y/Z fields in all the load commands. But since there're no users
of the fields except od, we can just add a function that properly prints
all version components out of 4-byte unsigned integer and remove the
version subfields until they're really needed.

This also fixes incorrect length of the first version subfield as it
should be two bytes long instead of one:
  X.Y.Z is encoded in nibbles xxxx.yy.zz

Signed-off-by: Roman Bolshakov <r.bolshakov@yadro.com>
Cc: Tristan Gingold <tgingold@free.fr>